+++ http4k-security/jwt/src/main/kotlin/org/http4k/security/jwt/JwtUtils.kt
@@ -1,0 +1,45 @@
+package org.http4k.security.jwt
+
+import com.nimbusds.jose.JWSAlgorithm
+import com.nimbusds.jose.jwk.JWK
+import com.nimbusds.jose.jwk.source.JWKSourceBuilder
+import com.nimbusds.jose.proc.JWSVerificationKeySelector
+import com.nimbusds.jose.proc.SecurityContext
+import com.nimbusds.jose.util.Resource
+import com.nimbusds.jose.util.ResourceRetriever
+import org.http4k.client.JavaHttpClient
+import org.http4k.core.HttpHandler
+import org.http4k.core.Method
+import org.http4k.core.Response
+import org.http4k.core.Status
+import org.http4k.core.Uri
+import org.http4k.lens.Header
+import java.io.IOException
+import java.net.URI
+
+fun <Context: SecurityContext> http4kJwsKeySelector(
+    jwkUri: Uri,
+    algorithm: JWSAlgorithm,
+    http: HttpHandler = JavaHttpClient()
+) = JWSVerificationKeySelector(
+    algorithm,
+    JWKSourceBuilder.create<Context>(
+        URI(jwkUri.toString()).toURL(),
+        http4kResourceRetriever(http)
+    ).build()
+)
+
+fun http4kResourceRetriever(http: HttpHandler) = ResourceRetriever { url ->
+    val response = org.http4k.core.Request(Method.GET, url.toString()).let(http)
+    if (!response.status.successful) throw IOException("Error retrieving JWK from $url: $response")
+    Resource(response.bodyString(), Header.CONTENT_TYPE(response)?.value)
+}
+
+fun jwkServer(publicKeys: Collection<JWK>): HttpHandler {
+    val response = Response(Status.OK)
+        .body( """{"keys":[${publicKeys.joinToString(",")}]}""")
+        .header("Content-Type", "application/json")
+    return { response }
+}
+
+fun jwkServer(vararg publicKeys: JWK) = jwkServer(publicKeys.toList())