+++ http4k-security/jwt/src/test/kotlin/org/http4k/security/jwt/JwtAuthorizerTest.kt
@@ -1,0 +1,176 @@
+package org.http4k.security.jwt
+
+import com.natpryce.hamkrest.absent
+import com.natpryce.hamkrest.assertion.assertThat
+import com.natpryce.hamkrest.equalTo
+import com.natpryce.hamkrest.present
+import com.nimbusds.jose.JWSAlgorithm
+import com.nimbusds.jose.jwk.KeyUse
+import com.nimbusds.jose.jwk.RSAKey
+import com.nimbusds.jose.proc.SingleKeyJWSKeySelector
+import com.nimbusds.jwt.JWTClaimsSet
+import org.http4k.core.Response
+import org.http4k.core.Status.Companion.NOT_FOUND
+import org.http4k.core.Uri
+import org.junit.jupiter.api.Test
+import java.time.Clock
+import java.time.Duration
+import java.time.Instant
+import java.time.ZoneId
+import java.time.ZoneOffset
+
+class JwtAuthorizerTest {
+    private val rsa = RsaProvider("testServer")
+
+    @Test
+    fun `process invalid jwt`() {
+        val provider = JwtAuthorizer(
+            keySelector = SingleKeyJWSKeySelector(JWSAlgorithm.RS256, rsa.publicKey),
+            lookup = { it.subject }
+        )
+        assertThat(provider("lolcats"), absent())
+    }
+
+    @Test
+    fun `process valid jwt for unauthorized user`() {
+        val token = rsa.generate("subject")
+        val provider = JwtAuthorizer(
+            keySelector = SingleKeyJWSKeySelector(JWSAlgorithm.RS256, rsa.publicKey),
+            lookup = { null }
+        )
+        assertThat(provider(token), absent())
+    }
+
+    @Test
+    fun `process local HS jwt`() {
+        val hs = HsProvider( "testServer")
+        val token = hs.generate("sub1")
+        assertThat(hs.verify(token), equalTo("sub1"))
+    }
+
+    @Test
+    fun `get verified subject`() {
+        val token = rsa.generate("sub1")
+        val provider = JwtAuthorizer(
+            keySelector = SingleKeyJWSKeySelector(JWSAlgorithm.RS256, rsa.publicKey),
+            lookup = { it.subject }
+        )
+
+        assertThat(provider(token), equalTo("sub1"))
+    }
+
+    @Test
+    fun `process remote jwt`() {
+        val token = rsa.generate("sub1")
+        val provider = JwtAuthorizer(
+            lookup =  { it.subject },
+            keySelector = http4kJwsKeySelector(
+                jwkUri = Uri.of("http://localhost/keys.jwks"),
+                algorithm = JWSAlgorithm.RS256,
+                http = jwkServer(
+                    RSAKey.Builder(rsa.publicKey)
+                        .keyUse(KeyUse.SIGNATURE)
+                        .keyID("key1")
+                        .build()
+                        .toPublicJWK()
+                )
+            )
+        )
+
+        assertThat(provider(token), equalTo("sub1"))
+    }
+
+    @Test
+    fun `process remote jwt - 404`() {
+        val provider = JwtAuthorizer(
+            lookup = { it.subject },
+            keySelector = http4kJwsKeySelector(
+                jwkUri = Uri.of("http://localhost/keys.jwks"),
+                algorithm = JWSAlgorithm.RS256,
+                http = { Response(NOT_FOUND) }
+            )
+        )
+
+        assertThat(provider(rsa.generate("sub1")), absent())
+    }
+
+    @Test
+    fun `verify issuer`() {
+        val provider = JwtAuthorizer(
+            lookup = { it.subject },
+            keySelector = SingleKeyJWSKeySelector(JWSAlgorithm.RS256, rsa.publicKey),
+            exactMatchClaims = JWTClaimsSet.Builder()
+                .issuer(rsa.issuer)
+                .build()
+        )
+
+        val token = rsa.generate("sub1")
+        assertThat(provider(token), equalTo("sub1"))
+    }
+
+    @Test
+    fun `verify audience`() {
+        val provider = JwtAuthorizer(
+            keySelector = SingleKeyJWSKeySelector(JWSAlgorithm.RS256, rsa.publicKey),
+            audience = setOf("foo", "bar"),
+            lookup = { it.subject }
+        )
+
+        assertThat(rsa.generate("sub1", audience = emptyList()).let(provider), absent())
+        assertThat(rsa.generate("sub1", audience = listOf("foo")).let(provider), present())
+        assertThat(rsa.generate("sub1", audience = listOf("foo", "bar")).let(provider), present())
+        assertThat(rsa.generate("sub1", audience = listOf("foo", "bar", "baz")).let(provider), present())
+        assertThat(rsa.generate("sub1", audience = listOf("baz")).let(provider), absent())
+    }
+
+    @Test
+    fun `verify issuer - invalid`() {
+        val provider = JwtAuthorizer(
+            lookup = { it.subject },
+            keySelector = SingleKeyJWSKeySelector(JWSAlgorithm.RS256, rsa.publicKey),
+            exactMatchClaims = JWTClaimsSet.Builder()
+                .issuer("issuer2")
+                .build()
+        )
+
+        val token = rsa.generate("sub1")
+        assertThat(provider(token), absent())
+    }
+
+    @Test
+    fun `verify extra claims`() {
+        val provider = JwtAuthorizer(
+            keySelector = SingleKeyJWSKeySelector(JWSAlgorithm.RS256, rsa.publicKey),
+            exactMatchClaims = JWTClaimsSet.Builder()
+                .claim("foo", "1")
+                .build(),
+            requiredClaims = setOf("bar"),
+            prohibitedClaims = setOf("baz"),
+            lookup = { it.subject }
+        )
+
+        assertThat(provider(rsa.generate("sub1")), absent())
+        assertThat(provider(rsa.generate("sub1", "foo" to "1")), absent())
+        assertThat(provider(rsa.generate("sub1", "foo" to "2", "bar" to "1")), absent())
+        assertThat(provider(rsa.generate("sub1", "foo" to "1", "bar" to "1")), present())
+        assertThat(provider(rsa.generate("sub1", "foo" to "1", "bar" to "1", "baz" to "1")), absent())
+    }
+
+    @Test
+    fun `verify expiry`() {
+        val clock = object: Clock() {
+            override fun instant() = Instant.parse("2024-01-07T12:00:00Z")
+            override fun withZone(zone: ZoneId?) = TODO()
+            override fun getZone() = ZoneOffset.UTC
+        }
+
+        val provider = JwtAuthorizer(
+            keySelector = SingleKeyJWSKeySelector(JWSAlgorithm.RS256, rsa.publicKey),
+            lookup = { it.subject },
+            clock = clock
+        )
+
+        assertThat(rsa.generate("sub1", expires = clock.instant() - Duration.ofMinutes(1)).let(provider), absent())
+        assertThat(rsa.generate("sub1", expires = clock.instant() + Duration.ofMinutes(1)).let(provider), present())
+    }
+}